move ActivityActinos into compositeBarActions.ts
